requirements.txt
Django==1.11.1
djangorestframework==3.6.3
django-filter==1.0.2
djangorestframework-xml==1.3.0
django-reversion==2.0.8
